Move the functions converted to ts to virtual-table.ts.
